backend/app/visualization_exporter.py
# app/visualization_exporter.py
from PIL import Image
import io
import subprocess
import tempfile

def generate_preview_from_gltf(gltf_bytes: bytes, width=512, height=512) -> bytes:
    """
    Simple strategy: use external headless glTF renderer (e.g., Babylon headless or blender script) to render preview.
    For a lightweight demo, render a placeholder icon with title text.
    """
    img = Image.new("RGBA", (width, height), (11,18,32,255))
    # draw placeholder
    # TODO: Add actual rendering logic using a glTF renderer
    # For now, return a simple placeholder image
    return img.tobytes()


